Changed query method to use query-all, thus fixing every problem at once
query-all is a secret undocumented command that produces a complete query list, containing all the individual queries that are documented plus many more. This includes disc number and count queries that actually work, and an artwork-id query that obviates the need for us to attempt to build the MD5 hash and corresponding ID ourselves. The script is now much more simple and reliable

Signed-off-by: Rhotias <rhotias@hotmail.com>